Parse `assert` and `assert_eq` as expressions
