+++ compiler/noirc_frontend/src/hir/comptime/display.rs
@@ -573,6 +573,12 @@ fn remove_interned_in_expression_kind(
                 ..*call
             }))
         }
+        ExpressionKind::Constrain(constrain) => ExpressionKind::Constrain(ConstrainExpression {
+            arguments: vecmap(constrain.arguments, |expr| {
+                remove_interned_in_expression(interner, expr)
+            }),
+            ..constrain
+        }),
         ExpressionKind::Constructor(constructor) => {
             ExpressionKind::Constructor(Box::new(ConstructorExpression {
                 fields: vecmap(constructor.fields, |(name, expr)| {
@@ -727,12 +733,6 @@ fn remove_interned_in_statement_kind(
             expression: remove_interned_in_expression(interner, let_statement.expression),
             r#type: remove_interned_in_unresolved_type(interner, let_statement.r#type),
             ..let_statement
-        }),
-        StatementKind::Constrain(constrain) => StatementKind::Constrain(ConstrainExpression {
-            arguments: vecmap(constrain.arguments, |expr| {
-                remove_interned_in_expression(interner, expr)
-            }),
-            ..constrain
         }),
         StatementKind::Expression(expr) => {
             StatementKind::Expression(remove_interned_in_expression(interner, expr))
